# docs/openapi.yaml
openapi: 3.0.3
info:
  title: assignP API
  version: 1.0.0
  description: REST API for the assignP task management application.
servers:
  - url: /api
paths:
  /auth/login:
    post:
      summary: Authenticate and receive a JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuth'
        '401':
          description: Invalid credentials
  /auth/register:
    post:
      summary: Register a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
  /auth/logout:
    post:
      summary: Logout the current session and invalidate the JWT.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
  /status/stream:
    get:
      summary: Real-time server status stream.
      responses:
        '200':
          description: Server-sent events stream
          content:
            text/event-stream: {}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    ApiResponseAuth:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            token:
              type: string
            role:
              type: string
    ApiResponseUser:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserDTO'
    ApiResponseMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string